requirements.txt
pyspark
kafka-python
tweepy
textblob
streamlit
pandas
matplotlib
wordcloud
nltk
